workflows/requirements.txt
polygon-api-client
polars
pandera[polars]
flytekitplugins-pandera
